Split branding module into focused submodules
